Modified with protocol and oops concept.
